requirements.txt
psycopg2-binary==2.9.9
setuptools==69.5.1
unicode==2.9
Unidecode==1.3.8
wheel==0.43.0
